Added helper functions to identify if strings are numbers in order for the parser and tokenizer to make nodes that the syntax tree generator can interpret as numbers.
