+++ src/token.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "states.h"
 #include <stdio.h>
+#include <string.h>
 #define MAX_TOKEN_SIZE 100
 
 #define TYPE_TOKEN_LPAREN STATE_LIST_START
@@ -24,21 +25,6 @@ typedef struct _token token;
 token * tokenAlloc(int numTokens){
 	return (token *) malloc(numTokens * sizeof(token));
 }
-/*
-token * makeToken(int state, int lineNumber, char* text){
-	token * out = tokenAlloc(1);
-	out -> state = state;
-	out -> lineNumber = lineNumber;
-	out -> text = (char * ) malloc(sizeof(char) * (strlen(text) + 1));
-	strcpy(out -> text, text);
-	return out;
-}
-
-void freeToken(token * t){
-	free (t->text);
-	free (t);
-}
-*/
 
 /**
  * Performs a shallow copy of the fields of the token pointed to by ptr
@@ -67,7 +53,7 @@ void fgetToken(FILE *file, token * container){
 void getToken(token * container){
 	fgetToken(stdin, container);
 }
-
+//TODO: change to token pointer to conserve memory and speed compilation time
 void fprintToken(FILE* file, token t){
 	fprintf(file, "%d %d %d %s\n",t.type, t.lineNumber, t.colNumber, t.text);
 }
@@ -91,4 +77,73 @@ void printTokenData(int state, int lineNumber, int colNumber, char * text){
 
 void printTokenDebug(token t){
 	printf("{token type:%d, line:%d, col:%d, text:\"%s\"}",t.type, t.lineNumber, t.colNumber, t.text);
+}
+
+int is_id_start(char c){
+	return isalpha(c) || (c == '_');
+}
+
+int is_digit(char c){
+	return isdigit(c);
+}
+
+int is_id(char c){
+	return is_id_start(c) || is_digit(c) || c=='?' || c=='!' || c=='-' || c=='<' || c=='>';
+}
+
+/**
+ * Returns true if c is a whitespace character
+ */
+int is_whitespace(char c){
+	return isspace(c);
+}
+
+/**
+ * Returns true if the character is considered an operator.
+ * Valid operators are + - * / %  = & | < > !
+ */
+int is_op_char(char c){
+	return c=='+' || c=='-' || c=='*' || c=='/' || c=='%' || c=='=' || c=='&' || c=='|' || c== '<' || c=='>' || c=='!';
+}
+
+/**
+ * Returns true if the character is considered a brace -
+ * that is, if the character is a parenthesis, curly brace, or square brace
+ */
+int is_brace(char c){
+	return c=='(' || c==')' || c=='{' || c=='}' || c=='[' || c==']';
+}
+
+/**
+ * Returns true if the cstring s represents a valid integer in the language, and false otherwise.
+ * Currently, s is an integer if it contains only numeric digits.
+ */
+int is_integer(char * s){
+	int len = strlen(s);
+	for(int i = 0 ; i < len ; i++){
+		if(!is_digit(s[i]))
+			return false;
+	}
+	return true;
+}
+
+/**
+ * Returns true if the cstring s represents a valid float in the language, and false otherwise.
+ * Currently, s is an float if it contains only numeric digits and a single, optional decimal point.
+ * All integers are valid floats.
+ */
+int is_float(char * s){
+	int hasFoundDecimalPoint = 0;
+	int len = strlen(s);
+	for(int i = 0 ; i < len ; i++){
+		if((s[i] == '.'){
+			if(hasFoundDecimalPoint)
+				return false;
+			else
+				hasFoundDecimalPoint = true;
+		} else if(!is_digit(s[i])){
+			return false;
+		}
+	}
+	return true;
 }
